crm: drag-n-drop styling
