+++ src/packages/frontend/_antd_fix.sass
@@ -24,9 +24,6 @@ html
 
 
 .ant-popover-content
-  border: 1px solid $COL_GRAY_L
-  border-radius: 3px
-
   .ant-popover-arrow
     border-color: $COL_GRAY_L !important
     z-index: -1 !important
