+++ src/packages/frontend/frame-editors/crm-editor/views/view-menu/limit.tsx
@@ -1,7 +1,12 @@
 import { InputNumber, Space } from "antd";
 import { Icon } from "@cocalc/frontend/components";
 
-export default function limitMenu({ limit, setLimit }) {
+export default function limitMenu({
+  limit,
+  setLimit,
+  setRecordHeight,
+  recordHeight,
+}) {
   return {
     label: <span style={{ padding: "5px" }}>Limit ({limit})</span>,
     key: "limit",
@@ -11,6 +16,16 @@ export default function limitMenu({ limit, setLimit }) {
         disabled: true,
         label: <Limit limit={limit} setLimit={setLimit} />,
         key: "the-limit",
+      },
+      {
+        label: (
+          <Height
+            setRecordHeight={setRecordHeight}
+            recordHeight={recordHeight}
+          />
+        ),
+        key: "height",
+        disabled: true,
       },
     ],
   };
@@ -30,6 +45,27 @@ function Limit({ limit, setLimit }) {
           if (value) {
             setLimit(value);
           }
+        }}
+      />
+    </Space>
+  );
+}
+
+function Height({ setRecordHeight, recordHeight }) {
+  return (
+    <Space>
+      <div style={{ color: "#666" }}>Max height of records:</div>
+      <InputNumber
+        placeholder="Height..."
+        onChange={(value) => setRecordHeight(value)}
+        value={recordHeight}
+        min={20}
+        max={3000}
+        step={50}
+        style={{
+          width: "100px",
+          marginLeft: "5px",
+          marginBottom: "5px",
         }}
       />
     </Space>
